Remove prev TURBO and update tests to use new class
